extract interest rate logics to a separate library
