+++ contracts/red-bank/tests/test_liquidate.rs
@@ -14,15 +14,15 @@ use helpers::{
     th_get_expected_indices, th_get_expected_indices_and_rates, th_init_market, th_setup,
     TestUtilizationDeltaInfo,
 };
+use mars_interest_rate::{
+    compute_scaled_amount, compute_underlying_amount, get_scaled_liquidity_amount,
+    ScalingOperation, SCALING_FACTOR,
+};
 use mars_params::types::AssetParams;
 use mars_red_bank::{
     contract::execute,
     error::ContractError,
     execute::liquidation_compute_amounts,
-    interest_rates::{
-        compute_scaled_amount, compute_underlying_amount, get_scaled_liquidity_amount,
-        ScalingOperation, SCALING_FACTOR,
-    },
     state::{COLLATERALS, DEBTS, MARKETS},
 };
 use mars_red_bank_types::{
