+++ contracts/red-bank/tests/test_query.rs
@@ -1,7 +1,7 @@
 use cosmwasm_std::{testing::mock_env, Addr, Decimal, Uint128};
 use helpers::{set_collateral, th_init_market, th_setup};
+use mars_interest_rate::{get_scaled_debt_amount, get_underlying_debt_amount, SCALING_FACTOR};
 use mars_red_bank::{
-    interest_rates::{get_scaled_debt_amount, get_underlying_debt_amount, SCALING_FACTOR},
     query::{query_user_collaterals, query_user_debt, query_user_debts},
     state::DEBTS,
 };
