+++ packages/interest-rate/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name          = "mars-interest-rate"
+description   = "Computations related to interest rates"
+version       = { workspace = true }
+authors       = { workspace = true }
+edition       = { workspace = true }
+license       = { workspace = true }
+repository    = { workspace = true }
+homepage      = { workspace = true }
+documentation = { workspace = true }
+keywords      = { workspace = true }
+
+[lib]
+doctest = false
+
+[features]
+# for quicker tests, cargo test --lib
+# for more explicit tests, cargo test --features=backtraces
+backtraces = ["cosmwasm-std/backtraces"]
+
+[dependencies]
+cosmwasm-schema     = { workspace = true }
+cosmwasm-std        = { workspace = true }
+mars-red-bank-types = { workspace = true }
+mars-utils          = { workspace = true }
